# backend/docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/alerts/home:
    get:
      operationId: AlertsController_home
      responses:
        '200':
          description: Home SOS, status, recent activity, and Create Emergency Alert actions
      security:
        - access-token: []
      summary: Home for Manual Emergency Alert
      tags:
        - Alerts
  /api/alerts/modes:
    get:
      operationId: AlertsController_modes
      responses:
        '200':
          description: Emergency Mode vs Silent Mode
      security:
        - access-token: []
      summary: Alert Mode catalog
      tags:
        - Alerts
  /api/alerts/cancel-reasons:
    get:
      operationId: AlertsController_cancelReasons
      responses:
        '200':
          description: I'm Safe, False Alert, Test
      security:
        - access-token: []
      summary: Cancel Alert reasons
      tags:
        - Alerts
  /api/alerts/quick-responses:
    get:
      operationId: AlertsController_quickResponses
      responses:
        '200':
          description: Need Help, Send Location, I'm Safe
      security:
        - access-token: []
      summary: Live call quick responses
      tags:
        - Alerts
  /api/alerts/current:
    get:
      operationId: AlertsController_current
      responses:
        '200':
          description: Active alert or null
      security:
        - access-token: []
      summary: Current broadcasting alert
      tags:
        - Alerts
  /api/alerts/inbox:
    get:
      operationId: AlertsController_inbox
      parameters:
        - name: tab
          in: query
          required: false
          schema:
            type: string
            example: active
      responses:
        '200':
          description: Live banner, Active/Past tabs, invitations, MY GROUPS
      security:
        - access-token: []
      summary: Alerts tab — circle feed, past 7 days, group invitations
      tags:
        - Alerts
  /api/alerts/active:
    get:
      operationId: AlertsController_getActive
      parameters: []
      responses:
        '200':
          description: Broadcasting SOS alerts
        '403':
          description: Citizen JWT is not allowed
      security:
        - access-token: []
      summary: Admin — list broadcasting SOS alerts
      tags:
        - Alerts
  /api/alerts/{id}:
    get:
      operationId: AlertsController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            example: alt-active-991
            type: string
      responses:
        '200':
          description: ''
        '404':
          description: Alert not found
      security:
        - access-token: []
      summary: Get one alert by id
      tags:
        - Alerts
  /api/alerts/trigger:
    post:
      operationId: AlertsController_trigger
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerAlertDto'
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      summary: Create Emergency Alert / Quick / SOS / Test
      tags:
        - Alerts
  /api/alerts/direct:
    post:
      operationId: AlertsController_direct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerAlertDto'
      responses:
        '201':
          description: Alert Sent — contacts notified, skip confirmation
      security:
        - access-token: []
      summary: Direct Emergency Alert (SOS hold or Quick Emergency)
      tags:
        - Alerts
  /api/alerts/{id}/telemetry:
    post:
      operationId: AlertsController_telemetry
      parameters:
        - name: id
          required: true
          in: path
          schema:
            example: alt-active-991
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryDto'
      responses:
        '201':
          description: ''
        '404':
          description: Alert not found
      security:
        - access-token: []
      summary: Push a GPS / battery telemetry point
      tags:
        - Alerts
  /api/alerts/{id}/resolve:
    post:
      operationId: AlertsController_resolve
      parameters:
        - name: id
          required: true
          in: path
          schema:
            example: alt-active-991
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveAlertDto'
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      summary: Resolve / cancel an alert (optional PIN)
      tags:
        - Alerts
...
  /api/alerts/{id}/call:
    get:
      operationId: AlertsController_call
      parameters:
        - name: id
          required: true
          in: path
          schema:
            example: alt-active-991
            type: string
      responses:
        '200':
          description: Zeegocloud room + token
      security:
        - access-token: []
      summary: Start Group Video Call
      tags:
        - Alerts
  /api/alerts/{id}/messages:
    get:
      operationId: AlertsController_messages
      parameters:
        - name: id
          required: true
          in: path
          schema:
            example: alt-active-991
            type: string
        - name: groupId
          required: false
          in: query
          schema:
            example: grp-family-01
            type: string
      responses:
        '200':
          description: Family thread, Online status, chat bubbles
      security:
        - access-token: []
      summary: Emergency group chat
      tags:
        - Alerts
    post:
      operationId: AlertsController_sendMessage
      parameters:
        - name: id
          required: true
          in: path
          schema:
            example: alt-active-991
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendAlertMessageDto'
      responses:
        '201':
          description: Message sent
      security:
        - access-token: []
      summary: Send a group chat message
      tags:
        - Alerts
  /api/alerts/{id}/respond:
    get:
      operationId: AlertsController_responderView
      parameters:
        - name: id
          required: true
          in: path
          schema:
            example: alt-active-991
            type: string
      responses:
        '200':
          description: LIVE ALERT with I'm Responding / Can't Help
        '403':
          description: Viewer is not in the owner's safety circle
        '404':
          description: Alert not found
      security:
        - access-token: []
      summary: LIVE ALERT responder view
      tags:
        - Alerts
    post:
      operationId: AlertsController_respond
      parameters:
        - name: id
          required: true
          in: path
          schema:
            example: alt-active-991
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RespondAlertDto'
      responses:
        '200':
          description: Responder status updated
        '403':
          description: Viewer is not in the owner's safety circle
      security:
        - access-token: []
      summary: I'm Responding or Can't Help
      tags:
        - Alerts
  /api/alerts/{id}/quick-response:
    post:
      operationId: AlertsController_quickResponse
      parameters:
        - name: id
          required: true
          in: path
          schema:
            example: alt-active-991
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuickResponseDto'
      responses:
        '200':
          description: Quick reply posted
      security:
        - access-token: []
      summary: Need Help / Send Location / I'm Safe
      tags:
        - Alerts
  /api/alerts/{id}/participants:
    post:
      operationId: AlertsController_participants
      parameters:
        - name: id
          required: true
          in: path
          schema:
            example: alt-active-991
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateParticipantDto'
      responses:
        '200':
          description: Participant updated
      security:
        - access-token: []
      summary: Update a call participant
      tags:
        - Alerts
  /api/alerts/{id}/cancel:
    post:
      operationId: AlertsController_cancel
      parameters:
        - name: id
          required: true
          in: path
          schema:
            example: alt-active-991
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveAlertDto'
      responses:
        '200':
          description: Alert Cancelled
      security:
        - access-token: []
      summary: Cancel Alert
      tags:
        - Alerts
...
info:
  title: SafeAlert Emergency API
  description: >-
    REST + Socket.io API for the SafeAlert emergency response system.


    **Demo accounts**

    - Super Admin: `admin@safealert.app` / `adminpassword`

    - Ops: `ops@safealert.app` / `opspassword`

    - Citizen (Sarah): `sarah.johnson@example.com` / PIN `1234`

    - Phone OTP for Sarah: `123456`


        Use **Authorize** after login and paste the JWT as a Bearer token for
        `/api/auth/me`.

        **Socket.io** (`ws://localhost:5002`, same origin as REST)

        - Client → server: `alert:join`, `alert:telemetry`, `alert:message:send`,
        `alert:quick_response`, `admin:subscribe`

        - Server → client: `alert:state`, `alert:telemetry:update`,
        `alert:messages:update`, `alert:resolved`, `notification:new`,
        `admin:alert:new`, `admin:alert:resolved`, `admin:alert:telemetry`,
        `admin:metrics`
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: JWT from POST /api/auth/login
  schemas:
...
    TriggerAlertDto:
      type: object
      properties:
        userId:
          type: string
          example: usr-sarah-101
        emergencyTypeId:
          type: string
          example: et-assault
        mode:
          enum:
            - EMERGENCY
            - SILENT
            - TEST
          type: string
          example: EMERGENCY
        latitude:
          type: number
          example: 40.712776
        longitude:
          type: number
          example: -74.005974
        address:
          type: string
          example: 123 Main St, New York, NY 10001
        source:
          enum:
            - MANUAL
            - QUICK
            - SOS
            - DIRECT
          type: string
          example: MANUAL
        alertAllGroups:
          type: boolean
          example: false
    TelemetryDto:
      type: object
      properties:
        latitude:
          type: number
          example: 40.71285
        longitude:
          type: number
          example: -74.00602
        speed:
          type: number
          minimum: 0
          example: 1.5
        heading:
          type: number
          example: 95
        accuracy:
          type: number
          example: 3
        batteryLevel:
          type: number
          example: 84
      required:
        - latitude
        - longitude
    ResolveAlertDto:
      type: object
      properties:
        reason:
          enum:
            - SAFE
            - FALSE_ALARM
            - TEST
          type: string
          example: SAFE
        notes:
          type: string
          example: Reached a safe location.
        pin:
          type: string
          example: '1234'
          description: User's 4-digit PIN
        userId:
          type: string
          example: usr-sarah-101
    QuickResponseDto:
      type: object
      properties:
        action:
          type: string
          example: need_help
      required:
        - action
    UpdateParticipantDto:
      type: object
      properties:
        participantId:
          type: string
        contactId:
          type: string
        status:
          enum:
            - CONNECTED
            - CALLING
          type: string
    SendAlertMessageDto:
      type: object
      properties:
        text:
          type: string
          example: I see your location, heading towards you now!
        groupId:
          type: string
          example: grp-family-01
      required:
        - text
    RespondAlertDto:
      type: object
      properties:
        action:
          type: string
          enum:
            - RESPONDING
            - CANT_HELP
          example: RESPONDING
      required:
        - action
...